Переносит настройки фотографий из отдельного файла в create-photo.js
