+++ js/create-photo.js
@@ -1,16 +1,45 @@
 import { getPhotoData } from './data.js';
-import { getPhotoSettings } from './photo-settings.js';
 import { getRandomUniqueNumbers, getRandomArrayElement, getRandomNumber } from './utils.js';
 
 const createPhoto = () => {
 
-  const generateCommentIds = getRandomUniqueNumbers(getPhotoSettings().CommentsIdQuantity.MIN, getPhotoSettings().CommentsIdQuantity.MAX);
-  const generagePtohoIds = getRandomUniqueNumbers(getPhotoSettings().PhotosQuantity.MIN, getPhotoSettings().PhotosQuantity.MAX);
-  const generatePhotoUrls = getRandomUniqueNumbers(getPhotoSettings().PhotosQuantity.MIN, getPhotoSettings().PhotosQuantity.MAX);
+  const PhotosQuantity = {
+    MIN: 1,
+    MAX: 25,
+  };
+
+  const AvatarsQuantity = {
+    MIN: 1,
+    MAX: 6,
+  };
+
+  const LikesQuantity = {
+    MIN: 15,
+    MAX: 200,
+  };
+
+  const CommentsIdQuantity = {
+    MIN: 1,
+    MAX: 1000,
+  };
+
+  const CommentsForPhotoQuantity = {
+    MIN: 0,
+    MAX: 30,
+  };
+
+  const MessagesForCommentQuantity = {
+    MIN: 1,
+    MAX: 2,
+  };
+
+  const generateCommentIds = getRandomUniqueNumbers(CommentsIdQuantity.MIN, CommentsIdQuantity.MAX);
+  const generagePtohoIds = getRandomUniqueNumbers(PhotosQuantity.MIN, PhotosQuantity.MAX);
+  const generatePhotoUrls = getRandomUniqueNumbers(PhotosQuantity.MIN, PhotosQuantity.MAX);
 
   const getCommentMessages = (messages) => {
     const array = [];
-    for (let i = 1; i <= getRandomNumber(getPhotoSettings().MessagesForCommentQuantity.MIN, getPhotoSettings().MessagesForCommentQuantity.MAX); i++) {
+    for (let i = 1; i <= getRandomNumber(MessagesForCommentQuantity.MIN, MessagesForCommentQuantity.MAX); i++) {
       array.push(getRandomArrayElement(messages));
     }
     return array;
@@ -18,7 +47,7 @@ const createPhoto = () => {
 
   const createComment = () => ({
     id: generateCommentIds(),
-    avatar: `img/avatar-${getRandomNumber(getPhotoSettings().AvatarsQuantity.MIN, getPhotoSettings().AvatarsQuantity.MAX)}.svg`,
+    avatar: `img/avatar-${getRandomNumber(AvatarsQuantity.MIN, AvatarsQuantity.MAX)}.svg`,
     message: getCommentMessages(getPhotoData().MESSAGES).join(' '),
     name: getRandomArrayElement(getPhotoData().NAMES),
   });
@@ -27,11 +56,9 @@ const createPhoto = () => {
     id: generagePtohoIds(),
     url: `photos/${generatePhotoUrls()}.jpg`,
     description: getPhotoData().DESCRIPTIONS[getRandomNumber(0, getPhotoData().DESCRIPTIONS.length - 1)],
-    likes: getRandomNumber(getPhotoSettings().LikesQuantity.MIN, getPhotoSettings().LikesQuantity.MAX),
-    comments: Array.from({length: getRandomNumber(getPhotoSettings().CommentsForPhotoQuantity.MIN, getPhotoSettings().CommentsForPhotoQuantity.MAX)}, createComment),
+    likes: getRandomNumber(LikesQuantity.MIN, LikesQuantity.MAX),
+    comments: Array.from({length: getRandomNumber(CommentsForPhotoQuantity.MIN, CommentsForPhotoQuantity.MAX)}, createComment),
   };
 };
 
 export { createPhoto };
-
-
