+++ js/main.js
@@ -1,4 +1,5 @@
-import { getPhotoSettings } from './photo-settings.js';
 import { createPhoto } from './create-photo.js';
 
-const photos = Array.from({length: getPhotoSettings().PhotosQuantity.MAX}, createPhoto);
+const photos = Array.from({length: createPhoto().PhotosQuantity.MAX}, createPhoto());
+
+console.log(photos);
